frontpage: split component definition and app rendering [+]
By decoupling the CommunitiesCardGroup definition and its ReactDOM.render, other
code can reuse the component independently.

- Also renames featured-communities to new-communities because that's waht it is.
