+++ invenio_communities/templates/semantic-ui/invenio_communities/frontpage.html
@@ -84,6 +84,6 @@
       </a>
   </div>
   <div class="ui divider hidden"></div>
-  <div id="featured-communities"></div>
+  <div id="new-communities"></div>
 </div>
 {%- endblock page_body %}
